+++ lib/src/qr.dart
@@ -8,6 +8,7 @@ import 'helpers/platform/configure_web.dart'
     if (dart.library.io) 'helpers/platform/configure_nonweb.dart';
 import 'helpers/widgets/stack_tree.dart';
 import 'overlays/qoverlay.dart';
+import 'routers/qdeclarative._router.dart';
 import 'routers/qrouter.dart';
 import 'routes/qroute.dart';
 import 'routes/qroute_children.dart';
@@ -91,17 +92,13 @@ class QRContext {
           navigator: navigator,
           addHistory: addHistory);
 
-  /// Add this routes as child for the route with name.
-  // void expandRoute(String name, List<QRoute> routes) {}
-  // Remove this route from the router
-  //void cleanRoute(String routerName, String routeName) {}
-
   /// return the current tree widget
   Widget getActiveTree() {
     return DebugStackTree(_manager.controllers);
   }
 
-  Future<T?> show<T>(QOverlay overlay) => overlay.show();
+  Future<T?> show<T>(QOverlay overlay, {String? name}) async =>
+      await overlay.show(name: name);
 
   /// create a controller to use with a Navigator
   QRouterController createRouterController(String name,
@@ -110,6 +107,10 @@ class QRContext {
           String? initPath,
           QRouteInternal? initRoute}) =>
       _manager.createController(name, routes, cRoutes, initPath, initRoute);
+
+  /// create a state to use with a declarative router
+  QDeclarativeController createDeclarativeRouterController(QKey key) =>
+      _manager.createDeclarativeRouterController(key);
 
   /// Navigate to this path.
   /// The package will try to get the right navigtor to this path.
@@ -161,6 +162,14 @@ class QRContext {
 
   /// try to pop the last active navigator or go to last path in the history
   bool back() {
+    // is proccesed by declerative
+    if (_manager.isDeclarative(QR.history.current.key.key)) {
+      final dCon = _manager.getDeclarative(QR.history.current.key.key);
+      if (dCon.pop()) {
+        return true;
+      }
+    }
+
     var lastNavi = QR.history.current.navigator;
     if (_manager.hasController(lastNavi)) {
       // Should navigator be removed? if the last path in history is the last